requirements.txt
# API y Web
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Procesamiento de datos
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2

# Almacenamiento
boto3==1.29.3
botocore==1.32.3

# Utilidades
python-dateutil==2.8.2
pytz==2023.3
cachetools==5.3.2
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging y monitoreo
structlog==23.2.0
python-json-logger==2.0.7 